adjust window filter to be 32bit
also update vcu motor temp to be 32bit

+++ Inc/window_filtering.h
@@ -13,17 +13,17 @@
 #include <stdlib.h>
 
 typedef struct window_filter {
-	uint16_t current_filtered;
-	uint16_t *prev_values;
+	uint32_t current_filtered;
+	uint32_t *prev_values;
 	uint32_t running_sum;
-	uint16_t current_idx;
+	uint32_t current_idx;
 	bool initialized;
 	size_t window_size;
 } window_filter_t;
 
-void window_filter_initialize(window_filter_t *filter, uint16_t initial_value, size_t window_size);
+void window_filter_initialize(window_filter_t *filter, uint32_t initial_value, size_t window_size);
 
-void window_filter_update(window_filter_t *filter, uint16_t new_value);
+void window_filter_update(window_filter_t *filter, uint32_t new_value);
 
 void window_filter_delete(window_filter_t *filter);
